// web/src/types/PropertyFilter.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { PropertyOperator } from "./PropertyOperator";
import type { PropertySource } from "./PropertySource";

export type PropertyFilter = { key: string, type: PropertySource, operator: PropertyOperator, 
/**
 * String, number, boolean, or an array of those. Ignored by
 * `is_set`/`is_not_set`.
 */
value: string | number | boolean | Array<string | number | boolean> | null, };